requirements.txt
agno
fastapi
uvicorn
pydantic
httpx
streamlit
python-dotenv
google-cloud-bigquery
google-auth